Add datetime synchronization setup step
On a fresh kuiper prebuilt image, the current date&time are not synchronized to anything and are usually long in the past. This breaks apt install and potentially other utilities as well. This setup step tries to first use NTP (which I found is not available on said fresh kuiper image) and alternatively queries a well-known HTTP server (in this case, google.com) for the current date&time, as a low precision but still OK fallback.

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>