GMSL717MIPI test
Because the AD-VIPER-SL board will not be connected to the internet,
this prepares all needed files (gmsl_lib, smbus2 python package, i2c
configurations) on the pi, then copies them over to the carrier.

A test is successful if the color band test pattern is received (tests
assume the connected pi cam v2 is functional) and matches the known
content hash of the test pattern. A failure would manifest as a failure
in the i2c programming phase, or a mismatching content hash for the
received frame.

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>